shafa_logic/requirements.txt
playwright
telethon
python-dotenv
inquirer
